Gate clippy in CI and clear the existing warnings
Add a dedicated clippy CI job running
`cargo clippy --workspace --all-targets -- -D warnings` (make lint-check),
so a new lint fails the build instead of accumulating unnoticed. The warning
count had drifted from 18 to about 37 precisely because nothing enforced it.

Clear every existing warning so the gate starts green:

- Mechanical fixes across the workspace: needless borrows, inlined format
  args, `assert!(!x)` over `assert_eq!(x, false)`, `slice::from_ref` over
  `&[x.clone()]`, `sort_by_key(Reverse(..))` over a hand-written reverse
  comparator, `if let` over a single-arm match, and similar.
- `l2_normalize` takes `&mut [f32]` instead of `&mut Vec<f32>`.
- The lock-file opens declare `.truncate(false)`: a lock file's bytes are
  never read or written, so opening it must not clobber an existing one.
- The gallery cache-prune loop folds two identical arms into one short-circuit
  that still never calls remove_file during a dry run.
- Removed a dead test-only helper and a duplicated test attribute.

Two subjective structural lints (type_complexity, too_many_arguments) are
allowed workspace-wide in the root manifest with a documented reason; the few
genuinely awkward cases (a matrix-index loop, test-fixture builders) carry a
justified allow at their site.

Update the CI notes in CLAUDE.md and add a changelog entry.

+++ crates/videre-core/src/image_decode.rs
@@ -113,7 +113,7 @@ mod tests {
         tiff.extend_from_slice(&0x0112u16.to_be_bytes()); // Orientation
         tiff.extend_from_slice(&3u16.to_be_bytes()); // SHORT
         tiff.extend_from_slice(&1u32.to_be_bytes()); // count
-        tiff.extend_from_slice(&(orientation as u16).to_be_bytes());
+        tiff.extend_from_slice(&orientation.to_be_bytes());
         tiff.extend_from_slice(&0u16.to_be_bytes()); // value padding
         tiff.extend_from_slice(&0u32.to_be_bytes()); // no next IFD
